// src/enums/key.enum.ts
export enum LocalStorageKeyEnums {
  ADDRESS_DATAS = 'address-datas',
  SESSION_DATAS = 'session-data',
}

export enum CsvFormattedKeyEnums {
  WALLET_PRIVATE_KEY = 'wallet_private_key',
  STARK_PRIVATE_KEY = 'stark_private_key',
  COLLECTION_ADDRESS = 'collection_address',
  TOKEN_ID = 'token_id',
  COLLECTION_ADDRESS2 = 'collection_address_2',
  TOKEN_ID2 = 'token_id_2',
  WALLET_NAME = 'wallet_name',
  TARGET_WALLET = 'target_wallet',
  ADDRESS = 'address',
}